py-server-app/pyproject.toml
[tool.poetry]
name = "py-server-app"
version = "0.1.0"
description = ""
authors = ["Marcus <marc303i@stud.kea.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
xmltodict = "^0.14.2"
pandas = "^2.2.3"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
